Added account and refactored
Account is the top level user abstraction - a single address with public
and private keys available.

Renamed package to reflect this change.
