+++ go.mod
@@ -5,8 +5,8 @@ replace github.com/safex/gosafex => ./
 require (
 	github.com/BurntSushi/toml v0.3.1 // indirect
 	github.com/ebfe/keccak v0.0.0-20150115210727-5cc570678d1b
-	github.com/google/pprof v0.0.0-20190108002400-985cf9b4fdd2 // indirect
 	github.com/golang/protobuf v1.2.0
+	github.com/google/pprof v0.0.0-20190108002400-985cf9b4fdd2 // indirect
 	github.com/inconshreveable/mousetrap v1.0.0 // indirect
 	github.com/onsi/ginkgo v1.7.0 // indirect
 	github.com/onsi/gomega v1.4.3 // indirect
