+++ pkg/account/address_test.go
@@ -1,4 +1,4 @@
-package address
+package account
 
 import (
 	"bytes"
