+++ pkg/account/network_id.go
@@ -1,4 +1,4 @@
-package address
+package account
 
 import (
 	"encoding/binary"
